Controller setup - add schemas based on CRDs available
Since KServe controllers are modified to watch resources based on
available CRDs, a similar change in the setup of the manager is needed:
schemas need to be added to the manager based on available CRDs rather
than based only on the values in the inferenceservice-config ConfigMap.
This would keep both manager setup and controller setup in sync with
regards schemas and watches around the CRDs.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>